requirements.txt
anthropic==0.40.0
psycopg2-binary==2.9.9
pgvector==0.3.2
pdfplumber==0.11.0
python-dotenv==1.0.1
rank-bm25==0.2.2
tiktoken==0.7.0
pydantic==2.6.0
# Sprint 2
fastapi==0.111.0
uvicorn==0.29.0
streamlit==1.35.0
httpx==0.27.0
pytest-asyncio==0.23.6
voyageai
